extract uploadParts and remove debug log
